requirements.txt
Django==1.8.4
django-admin-bootstrapped==2.5.7
django-ckeditor-updated==4.4.4
django-mysql==1.0.1
mysqlclient==1.3.7
wheel==0.24.0
